moving deliveries map to components
