+++ components/delivery-address.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import deliveryAddressMap from '../helpers/delivery-address-map';
+import deliveryAddressMap from './delivery-address-map';
 import { identifyFTShippingZone, countriesSupportedISO } from '../helpers/supportedCountries';
 export function DeliveryAddress ({
 	fieldId = 'deliveryAddressFields',
